+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added <!-- for new features -->
 - Support for path-like objects as an option wherever filenames are passed in as arguments
+- Optional `encoding` parameter in `SubtitlesClip` [#1043]
 
 ### Changed <!-- for changes in existing functionality -->
 
@@ -35,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - When using `VideoClip.write_videofile()` with `write_logfile=True`, errors would not be properly reported [#890]
 - `TextClip.list("color")` now returns a list of bytes, not strings [#1119]
 - `TextClip.search("colorname", "color")` does not crash with a TypeError [#1119]
+- Using `rotate()` with a `ColorClip` no longer crashes [#1139]
 
 
 ## [v1.0.2](https://github.com/zulko/moviepy/tree/v1.0.2) (2020-03-26)